test(reader): split interaction policy coverage
Why:
- Reader gesture and fullscreen policy tests exercise BibleUI package contracts, not app-host bootstrap behavior.
- Keeping these tests in the app-host bundle continues paying the full app-build path for package-owned logic.

What Changed:
- Moved double-tap fullscreen gating, horizontal swipe mapping, and auto-fullscreen threshold tests into BibleUITests.
- Added behavior docblocks that describe the reader policy contracts and failure meaning.
- Updated the test architecture migration plan with the reader interaction-policy slice.

Validation:
- GitNexus impact checks for moved symbols reported LOW risk with no affected execution flows.
- GitNexus staged change detection reported no mapped production-flow changes.
- Ran focused BibleUITests ReaderInteractionPolicyTests package suite.
- Ran focused AndBibleTests reader-navigation remainder compile sentinel.
- Ran git diff checks, docblock guard, source-guards, and adversarial subagent review.

Impact:
- Reduces app-host unit-test surface while preserving reader gesture policy coverage in the package test lane.
- No production behavior changes.
